Refactor workspace/xreferences
- Move walkMostAST to own file
- Use Observables and generators
